Mention Errors on the Front End
Mention error explicitly on the front-end rather silently logging them
on server side

+++ src/generator/backend/dist.js
@@ -57,6 +57,8 @@ module.exports = {
   },
   cleanDist: function(appFolder, err) {
     fs.emptyDir(distPath + '/' + appFolder, (emptyErr) => {
+      if(emptyErr)
+         return err(emptyErr);
       fs.remove(distPath + '/' + appFolder, err);
     });
   },
